调整代理设置页面，对齐webdav。
